+++ src/Service/Lambda/src/Enum/LastUpdateStatusReasonCode.php
@@ -5,6 +5,7 @@ namespace AsyncAws\Lambda\Enum;
 final class LastUpdateStatusReasonCode
 {
     public const CAPACITY_PROVIDER_SCALING_LIMIT_EXCEEDED = 'CapacityProviderScalingLimitExceeded';
+    public const DEPENDENCY_ERROR = 'DependencyError';
     public const DISABLED_KMSKEY = 'DisabledKMSKey';
     public const DISALLOWED_BY_VPC_ENCRYPTION_CONTROL = 'DisallowedByVpcEncryptionControl';
     public const EC2REQUEST_LIMIT_EXCEEDED = 'EC2RequestLimitExceeded';
@@ -48,6 +49,7 @@ final class LastUpdateStatusReasonCode
     {
         return isset([
             self::CAPACITY_PROVIDER_SCALING_LIMIT_EXCEEDED => true,
+            self::DEPENDENCY_ERROR => true,
             self::DISABLED_KMSKEY => true,
             self::DISALLOWED_BY_VPC_ENCRYPTION_CONTROL => true,
             self::EC2REQUEST_LIMIT_EXCEEDED => true,