+++ src/Service/Lambda/src/ValueObject/FunctionConfiguration.php
@@ -300,7 +300,7 @@ final class FunctionConfiguration
     private $snapStart;
 
     /**
-     * The ARN of the runtime and any errors that occured.
+     * The ARN of the runtime and any errors that occurred.
      *
      * @var RuntimeVersionConfig|null
      */