+++ src/Service/MediaConvert/src/Enum/MpdC2paManifest.php
@@ -3,7 +3,7 @@
 namespace AsyncAws\MediaConvert\Enum;
 
 /**
- * When enabled, a C2PA compliant manifest will be generated, signed and embeded in the output. For more information on
+ * When enabled, a C2PA compliant manifest will be generated, signed and embedded in the output. For more information on
  * C2PA, see https://c2pa.org/specifications/specifications/2.1/index.html.
  */
 final class MpdC2paManifest