+++ src/Service/MediaConvert/src/ValueObject/CmfcSettings.php
@@ -83,7 +83,7 @@ final class CmfcSettings
     private $audioTrackType;
 
     /**
-     * When enabled, a C2PA compliant manifest will be generated, signed and embeded in the output. For more information on
+     * When enabled, a C2PA compliant manifest will be generated, signed and embedded in the output. For more information on
      * C2PA, see https://c2pa.org/specifications/specifications/2.1/index.html.
      *
      * @var CmfcC2paManifest::*|null