+++ src/Service/MediaConvert/src/ValueObject/Mp4Settings.php
@@ -32,7 +32,7 @@ final class Mp4Settings
     private $audioDuration;
 
     /**
-     * When enabled, a C2PA compliant manifest will be generated, signed and embeded in the output. For more information on
+     * When enabled, a C2PA compliant manifest will be generated, signed and embedded in the output. For more information on
      * C2PA, see https://c2pa.org/specifications/specifications/2.1/index.html.
      *
      * @var Mp4C2paManifest::*|null