+++ src/Service/ElastiCache/src/ValueObject/CacheCluster.php
@@ -240,7 +240,7 @@ final class CacheCluster
 
     /**
      *  If you are running Valkey or Redis OSS engine version 6.0 or later, set this parameter to yes if you want to opt-in
-     * to the next auto minor version upgrade campaign. This parameter is disabled for previous versions. .
+     * to the next auto minor version upgrade campaign. This parameter is disabled for previous versions.
      *
      * @var bool|null
      */