+++ src/Service/Lambda/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## NOT RELEASED
 
+### Fixed
+
+- Fix typos.
+
 ### Added
 
 - AWS api-change: AWS Lambda Durable Functions now supports customer managed KMS keys. This allows customers to configure a KMS key in Durable Config to have all their durable execution data encrypted.
@@ -65,7 +69,7 @@
 ### Added
 
 - AWS api-change: Rework regions configuration
-- AWS api-change: remove trailling `/` in `ListFunctions` request
+- AWS api-change: remove trailing `/` in `ListFunctions` request
 
 ### Changed